+++ changelog.md
@@ -3,6 +3,7 @@
 - Add ability pause and resume recordings
 - Add recording progress indicator
 - Remove playback on recording screen since it was not compatible with pausing of recording
+- Change default recording time to 24 seconds in preparation of future name change
 
 # 0.2.1
 - Fix issue on Listen page where # of recordings in the filter bar did not match the number in the